requirements.txt
notion-client==2.0.0
python-dotenv==1.0.0
schedule==1.2.1
absl-py==2.0.0
mypy==1.8.0
python-dateutil==2.8.2
isort==5.13.2
yapf==0.40.2
typing-extensions==4.14.0
parameterized==0.9.0
